+++ Voxel_raycaster/Voxel_raycaster/Renderers/TimePoint4DRenderer.cpp
@@ -10,7 +10,7 @@ TimePoint4DRenderer::~TimePoint4DRenderer()
 {
 }
 
-void TimePoint4DRenderer::Render(RenderContext const& rc, Tree4D const* tree, unsigned char* texture_array, float time_point) const
+void TimePoint4DRenderer::Render(RenderContext const& rc, Tree4D const* tree, unsigned char* texture_array, double time_point) const
 {
 	// Get the number of processors in this system
 	int iCPU = omp_get_num_procs();
@@ -18,9 +18,9 @@ void TimePoint4DRenderer::Render(RenderContext const& rc, Tree4D const* tree, un
 	// declare variables we use in loop
 	int x, index, partindex;
 	Tree4DTraverser t;
-	float maxNumberOfTimeSteps = tree->gridsize_T;
-	float unitlength_time = (tree->max[3] - tree->min[3])/ tree->gridsize_T;
-	float endtime = tree->max[3];
+	double maxNumberOfTimeSteps = tree->gridsize_T;
+	double unitlength_time = (tree->max[3] - tree->min[3])/ tree->gridsize_T;
+	double endtime = tree->max[3];
 
 #pragma omp parallel for private(x,t,v,index,factor)
 	for (int y = 0; y < rc.n_y; y++) {
@@ -30,8 +30,8 @@ void TimePoint4DRenderer::Render(RenderContext const& rc, Tree4D const* tree, un
 			Ray ray3D = rc.getRayForPixel(x, y);
 
 			// the number of time steps in which this voxels is filled.
-			float nbOfTimeSteps = 0.0;
-			for (float timestep = 0.0f; timestep < endtime; timestep = timestep + unitlength_time) {
+			double nbOfTimeSteps = 0.0;
+			for (double timestep = 0.0f; timestep < endtime; timestep = timestep + unitlength_time) {
 				Ray4D ray4D = Ray4D::convertRayTo4D(ray3D, timestep, 0.0f);
 				t = Tree4DTraverser(tree, ray4D);
 				while ((!t.isTerminated())) {
@@ -44,7 +44,7 @@ void TimePoint4DRenderer::Render(RenderContext const& rc, Tree4D const* tree, un
 					t.step();
 				}
 			}
-			float R, G, B;
+			double R, G, B;
 	
 			if (nbOfTimeSteps < 0.0 || nbOfTimeSteps > maxNumberOfTimeSteps)
 			{
@@ -68,14 +68,14 @@ void TimePoint4DRenderer::Render(RenderContext const& rc, Tree4D const* tree, un
 				G = 0.0f;
 				B = 0.0f;
 			}
-			float r = (255 * R);
-			float g = (255 * G);
-			float b = (255 * B);
+			double r = (255 * R);
+			double g = (255 * G);
+			double b = (255 * B);
 
 
-			/*	float r = 255;
-			float g = 0;
-			float b = 0;*/
+			/*	double r = 255;
+			double g = 0;
+			double b = 0;*/
 			texture_array[index] = (unsigned char)r;
 			texture_array[index + 1] = (unsigned char)g;
 			texture_array[index + 2] = (unsigned char)b;
